Peg project dependencies to specific versions.
Three days ago props began to fail in building when neotoma updated its master
branch to be 1.6, rather than 1.5. This commit ensures that all future builds
peg dependencies at known-to-build versions.

(An alternative would be to continue to target master and remove the 1.5
assertion, but I'm fond of building against a known target.)

Signed-off-by: Brian L. Troutwine <brian.troutwine@rackspace.com>