fixes setting array syntax
Existing handling of setting array indexes has a couple of issues. First, it calls a non-existent lists:append/3 in a couple
of places. It also mishandles setting the first element of an array greater than length one. This adds tests that cover the
problem spots in the code and fixes the problems.